requirements.txt
tqdm
colorama
pandas
click
pytest
seaborn
python-dateutil
numpy
pysam
deprecation
matplotlib
watchdog
psutil
pyfastx
braceexpand
networkx
pyfastx
watchdog
psutil
google-cloud-storage